+++ Combat/Ability/Service/Interface/IAbilitySlotCalculator.cs
@@ -4,6 +4,6 @@ namespace IdelPog.Combat.Ability.Service.Interface
 {
     public interface IAbilitySlotCalculator
     {
-        public byte GetAbilitySlots(EquippedAbility[] abilityStages);
+        public uint GetAbilitySlots(EquippedAbility[] abilityStages);
     }
 }
