+++ Combat/Assertion/Interface/IAbilityAssertion.cs
@@ -2,6 +2,6 @@
 {
     public interface IAbilityAssertion
     {
-        public void AssertAbilityCount(byte reservedAbilitySlots);
+        public void AssertAbilityCount(uint reservedAbilitySlots);
     }
 }
