+++ Combat/Exceptions/TooManyAbilitiesException.cs
@@ -4,7 +4,7 @@
     {
         private const string MESSAGE = "Oops! Too many abliities! You tried to equip {0} abilities when you set the MaxAbilities to {1}!!";
         
-        public TooManyAbilitiesException(byte abilitySlots, byte maxAbilities) : base(string.Format(MESSAGE, abilitySlots, maxAbilities))
+        public TooManyAbilitiesException(uint abilitySlots, uint maxAbilities) : base(string.Format(MESSAGE, abilitySlots, maxAbilities))
         {
         }
     }
